LayoutTests/canvas/philip/tests/2d.clearRect.globalalpha-expected.txt

< [index] >
2d.clearRect.globalalpha
clearRect is not affected by globalAlpha
References: 2d.clearRect
Actual output:
Expected output:

Passed
